Moved AdministrativePicking2 into the combat NS
That's where the client puts it in the skill window.
