+++ src/ChannelServer/ChannelServer.csproj
@@ -171,7 +171,7 @@
     <Compile Include="Skills\Hidden\BigBlessingWaterKit.cs" />
     <Compile Include="Skills\Hidden\Blessing.cs" />
     <Compile Include="Skills\Hidden\Dye.cs" />
-    <Compile Include="Skills\Life\AdministrativePicking2.cs" />
+    <Compile Include="Skills\Combat\AdministrativePicking2.cs" />
     <Compile Include="Skills\Life\FirstAid.cs" />
     <Compile Include="Skills\Life\Gathering.cs" />
     <Compile Include="Skills\Life\Herbalism.cs" />
